Some alerts added

+++ dashboard.php
@@ -95,29 +95,26 @@ if (!check_session()){
     <script>
 	function confirmation() {
 	    if (confirm("All virtual machines will be powered off and their initial snapshots recreated.\nProceed?")) {
-    	    // your deletion code
 		$('#populatealert').show();
 	     }
         return false;
 	}
-
 	function confirmation1() {
 	    if (confirm("All virtual machines will be powered off and their initial snapshots recreated.\nProceed?")) {
-    	    // your deletion code
 		$('#copyalert').show();
 	     }
         return false;
 	}
+	function confirmBox(text) {
+    	    return confirm(text);
+	}
     </script>
-    <script>
-</script>
   </head>
 <?php 
     require_once('functions/functions.php');
     $sql_reply=get_SQL_array("SELECT * FROM hypervisors");
 ?>
-  <body>
-
+<body>
 <!-- Modal vm info-->
 <div class="modal fade" id="vmInfo" tabindex="-1" role="dialog" aria-labelledby="vmInfo" aria-hidden="true">
     <div class="modal-dialog">
@@ -203,7 +200,6 @@ if (!check_session()){
 	<?php
 	    $x=0;
 	    while ($sql_reply[$x]['id']){
-//		$vms_query=get_SQL_array("SELECT * FROM vms WHERE hypervisor='{$sql_reply[$x][id]}' ORDER BY name");
 		$vms_query=get_SQL_array("SELECT vms.id,vms.name,vms.hypervisor,vms.machine_type,vms.source_volume,vms.snapshot,vms.maintenance,vms.filecopy,vms.state,vms_tmp.name AS sourcename  FROM vms LEFT JOIN vms AS vms_tmp ON vms.source_volume=vms_tmp.id WHERE vms.hypervisor='{$sql_reply[$x][id]}' ORDER BY vms.name");
 	?>
           <h1 class="sub-header"><?php echo "Hypervisor: " . $sql_reply[$x]['ip']; ?></h1>
@@ -283,9 +279,9 @@ if (!check_session()){
 			}
 			echo  '<a href="power.php?action=single&state=up&vm=' . $vms_query[$y]['id'] . '&hypervisor=' . $sql_reply[$x]['id'] . '" data-toggle="hover" class="btn ' . $pwr_button . '" aria-label="Power up" title="Power up">
 			      <span class="glyphicon glyphicon-play" aria-hidden="true"></span></a>
-			      <a href="power.php?action=single&state=down&vm=' . $vms_query[$y]['id'] . '&hypervisor=' . $sql_reply[$x]['id'] . '" data-toggle="hover" class="btn btn-default" aria-label="Shut down" title="Shut down (soft)">
+			      <a href="power.php?action=single&state=down&vm=' . $vms_query[$y]['id'] . '&hypervisor=' . $sql_reply[$x]['id'] . '" data-toggle="hover" class="btn btn-default" aria-label="Shut down" title="Shut down (soft)"  onclick="return confirmBox(' . "'Are you sure?'" . ');">
 			      <span class="glyphicon glyphicon-off" aria-hidden="true"></span></a>
-			      <a href="power.php?action=single&state=destroy&vm=' . $vms_query[$y]['id'] . '&hypervisor=' . $sql_reply[$x]['id'] . '" data-toggle="hover"  class="btn btn-danger" aria-label="Power down" title="Shut down (forced)">
+			      <a href="power.php?action=single&state=destroy&vm=' . $vms_query[$y]['id'] . '&hypervisor=' . $sql_reply[$x]['id'] . '" data-toggle="hover"  class="btn btn-danger" aria-label="Power down" title="Shut down (forced)"  onclick="return confirmBox(' . "'Are you sure?'" . ');">
 			      <span class="glyphicon glyphicon-remove-circle" aria-hidden="true"></span></a>';
 			      if ($pwr_status=="on"){
 				    echo' <a data-toggle="modal" data-target="#vmConsole" href="vm_screen.php?vm=' . $vms_query[$y]['id'] . '&hypervisor=' . $sql_reply[$x]['id'] . '" data-toggle="hover"  class="btn btn-info" aria-label="Open console" title="Open console">
